Remove custom TextInput builder
Allow using normal JDA components in our modals,
match using the custom IDs which JDA mandates
