+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/modals/IPartialModalData.kt
@@ -1,9 +1,6 @@
 package io.github.freya022.botcommands.internal.modals
 
-import gnu.trove.map.TLongObjectMap
-
 internal interface IPartialModalData {
     val handlerData: IModalHandlerData?
-    val inputDataMap: TLongObjectMap<InputData>
     val timeoutInfo: ModalTimeoutInfo?
 }
