--- BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/modals/InputData.kt
@@ -1,3 +1,0 @@
-package io.github.freya022.botcommands.internal.modals
-
-internal class InputData(val inputName: String)
