+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/modals/ModalBuilderImpl.kt
@@ -1,17 +1,10 @@
 package io.github.freya022.botcommands.internal.modals
 
-import dev.freya02.botcommands.jda.ktx.components.findAll
-import dev.freya02.botcommands.jda.ktx.components.toDefaultComponentTree
-import gnu.trove.map.TLongObjectMap
-import gnu.trove.map.hash.TLongObjectHashMap
 import io.github.freya022.botcommands.api.modals.Modal
 import io.github.freya022.botcommands.api.modals.ModalBuilder
 import io.github.freya022.botcommands.api.modals.ModalEvent
 import io.github.freya022.botcommands.api.modals.Modals
-import io.github.freya022.botcommands.internal.utils.classRef
 import io.github.freya022.botcommands.internal.utils.takeIfFinite
-import io.github.freya022.botcommands.internal.utils.throwState
-import net.dv8tion.jda.api.components.attribute.ICustomId
 import kotlin.time.Duration
 
 internal class ModalBuilderImpl internal constructor(
@@ -44,22 +37,8 @@ internal class ModalBuilderImpl internal constructor(
     }
 
     override fun build(): Modal {
-        //Extract input data into this map
-        val inputDataMap: TLongObjectMap<InputData> = TLongObjectHashMap()
-        components.toDefaultComponentTree()
-            .findAll<ICustomId>()
-            .forEach { actionComponent ->
-                val id = actionComponent.customId ?: return@forEach
-                val internalId = ModalMaps.parseInputId(id)
-
-                val data = modalMaps.consumeInput(internalId)
-                    ?: throwState("Modal component with id '$internalId' could not be found in the inputs created with the '${classRef<Modals>()}' class")
-                inputDataMap.put(internalId, data)
-            }
-
         internetSetId(modalMaps.insertModal(PartialModalData(
             handlerData,
-            inputDataMap,
             timeoutInfo ?: Modals.defaultTimeout.takeIfFinite()?.let { ModalTimeoutInfo(it, null) }
         )))
