+++ BotCommands-core/src/main/kotlin/io/github/freya022/botcommands/internal/modals/PartialModalData.kt
@@ -1,9 +1,6 @@
 package io.github.freya022.botcommands.internal.modals
 
-import gnu.trove.map.TLongObjectMap
-
 internal open class PartialModalData(
     override val handlerData: IModalHandlerData?,
-    override val inputDataMap: TLongObjectMap<InputData>,
     override val timeoutInfo: ModalTimeoutInfo?
 ) : IPartialModalData
