+++ BotCommands-core/src/test/kotlin/io/github/freya022/botcommands/test/commands/slash/SlashLocalization.kt
@@ -18,6 +18,7 @@ import io.github.freya022.botcommands.api.localization.context.localizeGuild
 import io.github.freya022.botcommands.api.localization.context.localizeUser
 import io.github.freya022.botcommands.api.localization.interaction.*
 import io.github.freya022.botcommands.api.modals.Modals
+import io.github.freya022.botcommands.api.modals.TextInput
 import io.github.freya022.botcommands.api.modals.annotations.RequiresModals
 import io.github.freya022.botcommands.api.modals.create
 import net.dv8tion.jda.api.components.textinput.TextInputStyle
